+++ changelog.md
@@ -1,3 +1,32 @@
+# v0.0.10
+- Added searching for common students in db client
+- Added finding common students functionality to commonstudents endpoint
+
+# v0.0.9
+- Added multiple registration for student/teacher in db client
+- Add register functionality to register endpoint
+
+# v0.0.8
+- Modified db functions to be more robust and use promises for registration of student and teacher
+
+# v0.0.7
+- Added mysql module for connection to mysql db
+- Added connection to db
+
+# v0.0.6
+- Added docker components to run locally
+- Added mysql and table schemas for teachers and students
+
+# v0.0.5
+- Setup retrievefornotifications endpoint
+- Modified error messages
+
+# v0.0.4
+- Setup suspend endpoint
+
+# v0.0.3
+- Setup commonstudents endpoint
+
 # v0.0.2
 - Setup register endpoint
 - Added Joi module for json validation